make the pages take the entire screen as desired
