+++ CowryWiseClone/Commons/Extensions/RootViewController+Extensions.swift
@@ -1,0 +1,35 @@
+//
+//  RootViewController+Extensions.swift
+//  CowryWiseClone
+//
+//  Created by waheedCodes on 02/10/2020.
+//  Copyright © 2020 waheedCodes. All rights reserved.
+//
+
+import UIKit
+
+extension RootViewController: UICollectionViewDelegate {
+    
+}
+
+extension RootViewController: UICollectionViewDataSource {
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return 4
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        
+        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: RootViewController.identifier, for: indexPath)
+        
+        cell.backgroundColor = .systemBlue
+        
+        return cell
+        
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath) -> CGSize {
+        return CGSize(width: view.frame.width, height: view.frame.height)
+    }
+    
+}
